Split protocols between various submodules
